net/phy: xgkr: Fixed BaseKR algorithm inconsistent coefficient updates
This is a corner case but when it occurs it affects the outcome of the BaseKR algorithm and the link training fails. In the case when coefficient are in the state 'NOT UPDATED' the status reported inside the algorithm is 'UPDATED' which is not consistent with the actual operation performed.

Signed-off-by: Florinel Iordache <florinel.iordache@nxp.com>