enetc: Fix hang with MDIO WA
The ENETC Tx clean-up flow ends up calling kmem_cache_free_bulk which
cannot be called with IRQs off.  The previous version of the MDIO WA
was keeping IRQs off over this call, this is now fixed.
Use a single lock in the enetc MSI handler and disable interrupts before
disabling preemption in MDIO accessors, following the model in the generic
code.

Fixes: 6907d6d1d1cc ("enetc: WA for MDIO register access issue")

Signed-off-by: Alex Marginean <alexandru.marginean@nxp.com>